requirements.txt
asgiref==3.5.2
Django==3.2.14
django-database-url==1.0.3
djangorestframework==3.13.1
mysqlclient==2.1.1
pytz==2022.1
sqlparse==0.4.2
dj-rest-auth==2.2.4
django-allauth==0.51.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0